Resolve conflicts in NEWS.md and R/utils.R:

- NEWS.md: keep the 6.0.1 history from main and add a new
  development-version section with an expanded description of the
  multi-error reporting and the new `collectErr` parameter on
  `.collectWarn()`.

- R/utils.R: preserve main's decision to export `.collectWarn` and
  add the `collectErr` parameter from the error-details branch. The
  roxygen documentation now describes the new parameter and the
  shape of the returned list.

+++ src/rprintf.c
@@ -21,7 +21,7 @@ int F77_SUB(basout)(int *io, int *lunit, char *string,long int nbcharacters){
              if we do basout(io,lunit,'     []',7) it works ...
              temp workaround , we returns to old version with a allocation
           */
-	  char *buffer = (char *)R_Calloc(nbcharacters+1,char);
+	  char *buffer = (char *)R_Calloc((size_t)nbcharacters+1,char);
           if (buffer) {
               strncpy(buffer,string,nbcharacters);
               buffer[nbcharacters]='\0';
